# src/swagger.json
{
  "info": {
      "description": "API for the PlagueCraft Generic Gambit",
      "title": "PCG-API",
      "version": "v1"
  },
  "openapi": "3.0.0",
  "paths": {
      "/bridges": {
          "get": {
              "parameters": [
                  {
                      "description": "",
                      "in": "query",
                      "name": "username",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful"
                  },
                  "404": {
                      "description": "User not found in database"
                  },
                  "500": {
                      "description": "Internal Server Error"
                  }
              },
              "summary": "Fetch game data for Bridges",
              "tags": [
                  "bridges"
              ]
          }
      },
      "/bridges/history": {
          "get": {
              "parameters": [
                  {
                      "description": "user to q for",
                      "in": "query",
                      "name": "user",
                      "required": false,
                      "schema": {
                          "type": "string"
                      }
                  },
                  {
                      "description": "how many objects you want returned",
                      "in": "query",
                      "name": "limit",
                      "required": false,
                      "schema": {
                          "type": "integer"
                      }
                  }
              ],
              "responses": {
                "200": {
                  "description": "Successful"
                },
                "404": {
                    "description": "User not found in database"
                },
                "500": {
                    "description": "Internal Server Error"
                }
              },
              "summary": "Fetch Bridges game history",
              "tags": [
                  "bridges"
              ]
          }
      },
      "/tntrun": {
          "get": {
              "parameters": [
                  {
                      "description": "",
                      "in": "query",
                      "name": "username",
                      "required": true,
                      "schema": {
                          "type": "string"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": "Successful"
                  },
                  "404": {
                      "description": "User not found in database"
                  },
                  "500": {
                      "description": "Internal Server Error"
                  }
              },
              "summary": "Fetch player data for TNTRUN",
              "tags": [
                  "tntrun"
              ]
          }
      },
      "/parkour": {
        "get": {
            "parameters": [
                {
                    "description": "",
                    "in": "query",
                    "name": "username",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                }
            ],
            "responses": {
                "200": {
                    "description": "Successful"
                },
                "404": {
                    "description": "User not found in database"
                },
                "500": {
                    "description": "Internal Server Error"
                }
            },
            "summary": "Fetch player data for Parkour Gambit",
            "tags": [
                "parkour"
            ]
        }
    },
      "/parkour/history": {
        "get": {
            "parameters": [
              {
                "description": "user to q for",
                "in": "query",
                "name": "user",
                "required": false,
                "schema": {
                    "type": "string"
                }
            },
            {
                "description": "how many objects you want returned",
                "in": "query",
                "name": "limit",
                "required": false,
                "schema": {
                    "type": "integer"
                }
            }
          ],
            "responses": {
                "200": {
                    "description": "Successful"
                },
                "500": {
                    "description": "Internal Server Error"
                }
            },
            "summary": "Fetch Parkour Gambit game history",
            "tags": [
                "parkour"
            ]
        }
    },
      "/tntrun/history": {
          "get": {
              "parameters": [
                {
                  "description": "user to q for",
                  "in": "query",
                  "name": "user",
                  "required": false,
                  "schema": {
                      "type": "string"
                  }
              },
              {
                  "description": "how many objects you want returned",
                  "in": "query",
                  "name": "limit",
                  "required": false,
                  "schema": {
                      "type": "integer"
                  }
              }
            ],
              "responses": {
                  "200": {
                      "description": "Successful"
                  },
                  "500": {
                      "description": "Internal Server Error"
                  }
              },
              "summary": "Fetch TNTRUN game history",
              "tags": [
                  "tntrun"
              ]
          }
      }
  },
  "servers": [
      {
          "url": "https://api.plaguecraft.xyz"
      }
  ],
  "tags": [
      {
          "description": "",
          "name": "tntrun"
      },
      {
          "description": "",
          "name": "bridges"
      },
      {
        "description": "",
        "name": "parkour"
      }
  ]
}